+++ oauth2/examples/outlook.rs
@@ -1,16 +1,16 @@
-use pimalaya_oauth2::{AuthorizationCodeGrant, RefreshAccessToken};
+use pimalaya_oauth2::{AuthorizationCodeGrant, Client, RefreshAccessToken};
 use std::env;
 
 pub fn main() {
     let port = env::var("PORT")
-        .unwrap_or(String::from("8080"))
+        .unwrap_or(String::from("9999"))
         .parse::<u16>()
         .expect("Invalid port.");
     let client_id = env::var("CLIENT_ID").expect("Missing the CLIENT_ID environment variable.");
     let client_secret =
         env::var("CLIENT_SECRET").expect("Missing the CLIENT_SECRET environment variable.");
 
-    let builder = AuthorizationCodeGrant::new(
+    let client = Client::new(
         client_id,
         client_secret,
         "https://login.microsoftonline.com/common/oauth2/v2.0/authorize",
@@ -18,20 +18,26 @@ pub fn main() {
     )
     .unwrap()
     .with_redirect_port(port)
-    .with_pkce()
-    // for managing emails
-    .with_scope("https://outlook.office.com/IMAP.AccessAsUser.All")
-    // for sending emails
-    .with_scope("https://outlook.office.com/SMTP.Send")
-    // for refresh token
-    .with_scope("offline_access");
+    .build()
+    .unwrap();
 
-    let client = builder.get_client().unwrap();
-    let (redirect_url, csrf_token) = builder.get_redirect_url(&client);
+    let auth_code_grant = AuthorizationCodeGrant::new()
+        .with_redirect_port(port)
+        .with_pkce()
+        // for managing emails
+        .with_scope("https://outlook.office.com/IMAP.AccessAsUser.All")
+        // for sending emails
+        .with_scope("https://outlook.office.com/SMTP.Send")
+        // for refresh token
+        .with_scope("offline_access");
+
+    let (redirect_url, csrf_token) = auth_code_grant.get_redirect_url(&client);
 
     println!("Go to: {}", redirect_url.to_string());
 
-    let (access_token, refresh_token) = builder.wait_for_redirection(&client, csrf_token).unwrap();
+    let (access_token, refresh_token) = auth_code_grant
+        .wait_for_redirection(&client, csrf_token)
+        .unwrap();
 
     println!("access token: {:?}", access_token);
     println!("refresh token: {:?}", refresh_token);