+++ oauth2/src/lib.rs
@@ -8,17 +8,24 @@
 //! ```
 
 pub mod authorization_code_grant;
+pub mod client;
 pub mod refresh_access_token;
 
 pub use authorization_code_grant::AuthorizationCodeGrant;
+pub use client::Client;
 pub use refresh_access_token::RefreshAccessToken;
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     #[error(transparent)]
-    AuthorizationCodeGrant(#[from] authorization_code_grant::Error),
+    ClientError(#[from] client::Error),
     #[error(transparent)]
-    RefreshAccessToken(#[from] refresh_access_token::Error),
+    AuthorizationCodeGrantError(#[from] authorization_code_grant::Error),
+    #[error(transparent)]
+    RefreshAccessTokenError(#[from] refresh_access_token::Error),
+
+    #[error(transparent)]
+    IoError(#[from] std::io::Error),
 }
 
 pub type Result<T> = std::result::Result<T, Error>;